# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: anime
    description: Anime data and streaming
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /anime/trending:
    get:
      operationId: getTrendingAnime
      tags: [anime]
      summary: Get top/trending anime
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 16
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/recent:
    get:
      operationId: getRecentAnime
      tags: [anime]
      summary: Get recently aired anime
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 16
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/search:
    get:
      operationId: searchAnime
      tags: [anime]
      summary: Search anime by title
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeListResponse"
  /anime/{malId}:
    get:
      operationId: getAnimeById
      tags: [anime]
      summary: Get anime details by MAL ID
      parameters:
        - name: malId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /anime/{malId}/episodes:
    get:
      operationId: getAnimeEpisodes
      tags: [anime]
      summary: Get episode list for an anime
      parameters:
        - name: malId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeListResponse"
  /anime/stream:
    get:
      operationId: getEpisodeStream
      tags: [anime]
      summary: Get streaming sources for an episode
      parameters:
        - name: episodeId
          in: query
          required: true
          schema:
            type: string
          description: Consumet episode ID (e.g. gogoanime episode slug)
        - name: provider
          in: query
          schema:
            type: string
            default: gogoanime
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [username, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, username, email, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
      required: [user, token]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    AnimeCard:
      type: object
      properties:
        malId:
          type: integer
        title:
          type: string
        image:
          type: string
        score:
          type: number
          nullable: true
        episodes:
          type: integer
          nullable: true
        status:
          type: string
          nullable: true
        genres:
          type: array
          items:
            type: string
        synopsis:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        type:
          type: string
          nullable: true
      required: [malId, title, image, genres]
    AnimeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AnimeCard"
        pagination:
          type: object
          properties:
            currentPage:
              type: integer
            hasNextPage:
              type: boolean
            total:
              type: integer
          required: [currentPage, hasNextPage, total]
      required: [data, pagination]
    AnimeDetail:
      type: object
      properties:
        malId:
          type: integer
        title:
          type: string
        titleEnglish:
          type: string
          nullable: true
        image:
          type: string
        trailer:
          type: string
          nullable: true
        score:
          type: number
          nullable: true
        scored_by:
          type: integer
          nullable: true
        rank:
          type: integer
          nullable: true
        episodes:
          type: integer
          nullable: true
        status:
          type: string
          nullable: true
        aired:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        rating:
          type: string
          nullable: true
        genres:
          type: array
          items:
            type: string
        synopsis:
          type: string
          nullable: true
        background:
          type: string
          nullable: true
        studios:
          type: array
          items:
            type: string
        type:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
      required: [malId, title, image, genres, studios]
    Episode:
      type: object
      properties:
        malId:
          type: integer
        number:
          type: integer
        title:
          type: string
          nullable: true
        aired:
          type: string
          nullable: true
        score:
          type: number
          nullable: true
        filler:
          type: boolean
        recap:
          type: boolean
      required: [malId, number, filler, recap]
    EpisodeListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
        total:
          type: integer
      required: [data, total]
    StreamSource:
      type: object
      properties:
        url:
          type: string
        quality:
          type: string
        isM3U8:
          type: boolean
      required: [url, quality, isM3U8]
    StreamResponse:
      type: object
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/StreamSource"
        headers:
          type: object
          additionalProperties:
            type: string
        episodeId:
          type: string
        provider:
          type: string
      required: [sources, episodeId, provider]